Conditions corrected

+++ src/main/java/school/mjc/stage0/conditions/task3/AliquotNumbers.java
@@ -2,8 +2,8 @@ package school.mjc.stage0.conditions.task3;
 
 public class AliquotNumbers {
     public void isFirstAliquot(int first, int second) {
-        int result = first % second;
-        if(result == 0){
+        float result = first % second;
+        if(result == 0.0f){
             System.out.println("Aliquot");
         }else{
             System.out.println("Not aliquot");
